fixed psf integration issue. cleaned pycache. added flow control

+++ PyPRIS/observe_station.py
@@ -85,7 +85,7 @@ class ObserveStation:
 
             if self.subpixel_shift is True:
                 # get the stamp with only the integer part
-                stamp = np.copy(self.psf[self.layerN, :, :])
+                stamp = np.float64(np.copy(self.psf[self.layerN, :, :]))
                 # up perform sub-pixel interpolation to include the sub-pixel shifts in the stamp.
 
                 locs1 = np.arange(0, stamp.shape[0])
